Rename: 테이블 컴포넌트 파일명 수정
