--- src/components/bithumb/index.ts
@@ -1,1 +1,0 @@
-export { Bithumb } from '../Table/Bithumb';
